NEW MAIN FILE: moved your work into testmapscript.py
if it's okay can we start using this file as our working file? sorry for the sudden switch but I decided reworking my part from the ground up would be easier than converting the mess of code in pythonfile.py; I've moved most of your stuff into this area and it should be ready to go!